Update int.sc encoding desc. for new say(-1) mode
I haven't actually ran the PHP to ensure that is valid but it seems
valid enough to me. I haven't actually seen the output of the change,
but again it seems valid enough to me. I don't think I could be missing
any subtleties and this change really is simple and doesn't warrant me
needing to install something that runs PHP just to check.